+++ packages/system/src/components/homepage/featured-libraries.tsx
@@ -1,5 +1,6 @@
 import classNames from "classnames"
 import React from "react"
+import { sampleSize } from "lodash"
 import {
 	featuredLibrariesStyle,
 	featuredLibrariesHeadingStyle,
@@ -8,8 +9,6 @@ import {
 } from "./featured-libraries.css"
 import { Library } from "../../models/library"
 import { LibraryCard } from "../cards/library-card"
-import { ComparisonBox } from "./comparison-box"
-import { getRandomNumberArray } from "./utils"
 
 export interface FeaturedLibrariesProps
 	extends React.ComponentPropsWithoutRef<"div"> {
@@ -24,30 +23,23 @@ export function FeaturedLibraries({
 	libraryTags,
 	...props
 }: FeaturedLibrariesProps) {
-	const indices = getRandomNumberArray(4, libraries.length)
-	const randomLibraries = indices.map((i) => libraries[i])
+	const randomLibraries = sampleSize(libraries, 6)
 	return (
 		<div className={classNames(className, featuredLibrariesStyle)} {...props}>
 			<h2 className={featuredLibrariesHeadingStyle}>
 				Featured React Libraries
 			</h2>
 			<div className={featuredLibrariesListStyle}>
-				<>
-					{randomLibraries.map((library, i) => (
-						<LibraryCard
-							key={`featured-library-${i}`}
-							library={library}
-							headingTag="h3"
-							onTagClick={() => {
-								return null
-							}}
-						/>
-					))}
-					<ComparisonBox
-						style={{ gridArea: "compare" }}
-						libraryTags={libraryTags}
+				{randomLibraries.map((library, i) => (
+					<LibraryCard
+						key={`featured-library-${i}`}
+						library={library}
+						headingTag="h3"
+						onTagClick={() => {
+							return null
+						}}
 					/>
-				</>
+				))}
 			</div>
 			<a href="/categories/libraries" className={featuredLibrariesViewAllStyle}>
 				View all libraries