+++ packages/system/src/sprinkles/sprinkles.css.ts
@@ -20,7 +20,7 @@ const space = {
 	...spaceInRem,
 }
 
-const borders = {
+export const borders = {
 	none: `none`,
 	thin: `1px solid ${vars.themeColors.outline}`,
 	light: `1px solid ${vars.themeColors.surfaceVariant}`,
@@ -271,8 +271,6 @@ const layout = defineProperties({
 })
 
 const colors = defineProperties({
-	conditions,
-	defaultCondition: "mobile",
 	properties: {
 		backgroundColor: vars.themeColors,
 		color: vars.themeColors,