+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+- [2022-01-05] Resolve known bugs created with merge of homepage content
+  [\#150](https://github.com/thisdot/framework.dev/pull/150)
+  ([tvanantwerp](https://github.com/tvanantwerp))
 - [2022-01-04] Fix trailing slashes on URLs
   [\#147](https://github.com/thisdot/framework.dev/pull/147)
   ([jbachhardie](https://github.com/jbachhardie))